refactor: rename rooms command
