+++ src/discord/commands/index.ts
@@ -1,9 +1,9 @@
 import Ping from './ping';
-import Rooms from './rooms';
 import Book from './book';
+import ManualBook from './manualbook';
 import Manage from './manage';
 import View from './view';
 // import Report from './report'; - temporarily disabled due to bug
 import Delete from './delete';
 
-export default [Rooms, Book, Manage, View, Delete, Ping];
+export default [Book, ManualBook, Manage, View, Delete, Ping];
